requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.1
MarkupSafe==3.0.2
openai==1.105.0
openpyxl==3.1.5
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xlrd==2.0.2
gunicorn==20.1.0
